CTCP-6450: Using presence of authToken to decide on whether to present sign out link.
