+++ a11y/templates/MainTemplateSpec.scala
@@ -27,10 +27,9 @@ class MainTemplateSpec extends A11ySpecBase {
 
     val title          = nonEmptyString.sample.value
     val timeoutEnabled = arbitrary[Boolean].sample.value
-    val canSignOut     = arbitrary[Boolean].sample.value
     val showBackLink   = arbitrary[Boolean].sample.value
 
-    val content = template.apply(title, timeoutEnabled, canSignOut, showBackLink) {
+    val content = template.apply(title, timeoutEnabled, showBackLink) {
       heading(title)
     }
