+++ app/views/SessionExpiredView.scala.html
@@ -17,18 +17,17 @@
 @import views.html.components.{Button, Heading}
 
 @this(
-        main_template: MainTemplate,
-        formHelper: FormWithCSRF,
-        heading: Heading,
-        button: Button
+    main_template: MainTemplate,
+    formHelper: FormWithCSRF,
+    heading: Heading,
+    button: Button
 )
 
 @()(implicit request: Request[?], messages: Messages)
 
 @main_template(
     title = messages("session_expired.title"),
-    showBackLink = false,
-    canSignOut = false
+    showBackLink = false
 ) {
 
     @formHelper(action = routes.SessionExpiredController.onSubmit(), Symbol("autoComplete") -> "off") {
