+++ app/views/templates/MainTemplate.scala.html
@@ -39,9 +39,10 @@
 @(
     title: String,
     timeoutEnabled: Boolean = true,
-    canSignOut: Boolean = true,
     showBackLink: Boolean = false
 )(mainContent: Html)(implicit request: Request[?], messages: Messages)
+
+@canSignOut = @{request.session.get("authToken").isDefined}
 
 @headScripts = {
     @if(canSignOut) {
@@ -71,7 +72,7 @@
     HmrcStandardPageParams(
         serviceURLs = ServiceURLs(
             serviceUrl = Some(appConfig.serviceUrl),
-            signOutUrl = if (canSignOut) Some(appConfig.signOutUrl) else None
+            signOutUrl = Option.when(canSignOut)(appConfig.signOutUrl)
         ),
         banners = Banners(
             phaseBanner = Some(betaBanner()),
